Fixed pattern scanning bug
Fixed a pattern scanning bug for data patterns where the address was not
being read properly. The result should be found by reading the address
at module base + offset pattern was found at + offset from the found
pattern the data is obtained, which it now accounts for.

Some minor documentation of interfaces added as well.